# call precompile since we know type info