Seems not like much of a time savings overall, but the info was there.