@SimonDanisch 2015-10-26
Interesting, that seems to work. I would have sworn, that this won't work, because at `precompile` time there is no active context. Either I missunderstand what `precompile` does, or this works by accidant, or I missunderstand this article: https://msdn.microsoft.com/en-us/library/windows/desktop/dd374386%28v=vs.85%29.aspx
My guess is, that even `precompile` doesn't generate native code yet, so it doesn't really do much with automatic precompilation turned on.