@SimonDanisch 2015-12-10
I was close to merging this, but I did some benchmarks first and the timings don't change a bit!
```julia
...setup...
function test()
    @time glViewport(0f0,0f0,Float32(wh),Float32(wh))
    @time glViewport(0f0,0f0,Float32(wh),Float32(wh))
    @time  glClearColor(0f0,0f0,0f0,0f0)
    @time  glClearColor(0f0,0f0,0f0,0f0)
    const result = GLuint[0]
    @time  glGenBuffers(GLint(1), result)
    @time  glGenBuffers(GLint(1), result)
end
test()
#with precompile
0.001207 seconds (180 allocations: 6.578 KB)
  0.000003 seconds (2 allocations: 32 bytes)
  0.000004 seconds
  0.000000 seconds
  0.000005 seconds
  0.000000 seconds


#without
 0.000693 seconds (180 allocations: 6.578 KB)
  0.000003 seconds (2 allocations: 32 bytes)
  0.000004 seconds
  0.000000 seconds
  0.000005 seconds
  0.000000 seconds
```
I guess this is because it just lowers to julia AST, and that is already done while precompiling?